pyproject.toml
[project]
name = "pivot-div-donchian-bot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "PyYAML>=6.0.1",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"
